- fix warnings
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@31350 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/ewl/src/lib/ewl_text.c
@@ -643,7 +643,7 @@ ewl_text_text_insert_private(Ewl_Text *t, const char *txt,
 	/* copy the text over, replace invalid UTF-8 chars */
 	tmp = (char *)txt;
 	ptr = t->text + bidx;
-	while (*tmp && (tmp - txt) < blen) 
+	while (*tmp && (tmp - txt) < (int)blen) 
 	{
 		if (ewl_text_char_utf8_is(tmp)) 
 		{
